[MIG] web_timeline: Migration to 16.0
Syntax changes

Syntax changes
