+++ web_timeline/__manifest__.py
@@ -4,7 +4,7 @@
 {
     "name": "Web timeline",
     "summary": "Interactive visualization chart to show events in time",
-    "version": "15.0.1.0.1",
+    "version": "16.0.1.0.0",
     "development_status": "Production/Stable",
     "author": "ACSONE SA/NV, "
     "Tecnativa, "
@@ -25,11 +25,9 @@
             "web_timeline/static/src/scss/web_timeline.scss",
             "web_timeline/static/src/js/timeline_view.js",
             "web_timeline/static/src/js/timeline_renderer.js",
-            "web_timeline/static/src/js/timeline_controller.js",
+            "web_timeline/static/src/js/timeline_controller.esm.js",
             "web_timeline/static/src/js/timeline_model.js",
             "web_timeline/static/src/js/timeline_canvas.js",
-        ],
-        "web.assets_qweb": [
             "web_timeline/static/src/xml/web_timeline.xml",
         ],
     },
