+++ web_timeline/i18n/web_timeline.pot
@@ -4,7 +4,7 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 15.0\n"
+"Project-Id-Version: Odoo Server 16.0\n"
 "Report-Msgid-Bugs-To: \n"
 "Last-Translator: \n"
 "Language-Team: \n"
@@ -14,42 +14,42 @@ msgstr ""
 "Plural-Forms: \n"
 
 #. module: web_timeline
-#. openerp-web
+#. odoo-javascript
 #: code:addons/web_timeline/static/src/js/timeline_renderer.js:0
 #, python-format
 msgid "<b>UNASSIGNED</b>"
 msgstr ""
 
 #. module: web_timeline
-#. openerp-web
-#: code:addons/web_timeline/static/src/js/timeline_controller.js:0
+#. odoo-javascript
+#: code:addons/web_timeline/static/src/js/timeline_controller.esm.js:0
 #, python-format
 msgid "Are you sure you want to delete this record?"
 msgstr ""
 
 #. module: web_timeline
-#. openerp-web
+#. odoo-javascript
 #: code:addons/web_timeline/static/src/xml/web_timeline.xml:0
 #, python-format
 msgid "Day"
 msgstr ""
 
 #. module: web_timeline
-#. openerp-web
+#. odoo-javascript
 #: code:addons/web_timeline/static/src/xml/web_timeline.xml:0
 #, python-format
 msgid "Month"
 msgstr ""
 
 #. module: web_timeline
-#. openerp-web
+#. odoo-javascript
 #: code:addons/web_timeline/static/src/js/timeline_renderer.js:0
 #, python-format
 msgid "Template \"timeline-item\" not present in timeline view definition."
 msgstr ""
 
 #. module: web_timeline
-#. openerp-web
+#. odoo-javascript
 #: code:addons/web_timeline/static/src/js/timeline_view.js:0
 #: model:ir.model.fields.selection,name:web_timeline.selection__ir_ui_view__type__timeline
 #, python-format
@@ -57,14 +57,14 @@ msgid "Timeline"
 msgstr ""
 
 #. module: web_timeline
-#. openerp-web
+#. odoo-javascript
 #: code:addons/web_timeline/static/src/js/timeline_renderer.js:0
 #, python-format
 msgid "Timeline view has not defined 'date_start' attribute."
 msgstr ""
 
 #. module: web_timeline
-#. openerp-web
+#. odoo-javascript
 #: code:addons/web_timeline/static/src/xml/web_timeline.xml:0
 #, python-format
 msgid "Today"
@@ -81,21 +81,21 @@ msgid "View Type"
 msgstr ""
 
 #. module: web_timeline
-#. openerp-web
-#: code:addons/web_timeline/static/src/js/timeline_controller.js:0
+#. odoo-javascript
+#: code:addons/web_timeline/static/src/js/timeline_controller.esm.js:0
 #, python-format
 msgid "Warning"
 msgstr ""
 
 #. module: web_timeline
-#. openerp-web
+#. odoo-javascript
 #: code:addons/web_timeline/static/src/xml/web_timeline.xml:0
 #, python-format
 msgid "Week"
 msgstr ""
 
 #. module: web_timeline
-#. openerp-web
+#. odoo-javascript
 #: code:addons/web_timeline/static/src/xml/web_timeline.xml:0
 #, python-format
 msgid "Year"
