+++ web_timeline/static/lib/vis-timeline/vis-timeline-graph2d.js
@@ -50109,4 +50109,3 @@
 	Object.defineProperty(exports, '__esModule', { value: true });
 
 }));
-//# sourceMappingURL=vis-timeline-graph2d.js.map
